+++ projects/core/src/page/title/facade/index.ts
@@ -1,5 +1,6 @@
 export * from './page-title.service';
-export * from './product-page-title.resolver';
-export * from './category-page-title.resolver';
-export * from './content-page-title.resolver';
-export * from './search-page-title.resolver';
+export * from './resolvers/page-title.resolver';
+export * from './resolvers/product-page-title.resolver';
+export * from './resolvers/category-page-title.resolver';
+export * from './resolvers/content-page-title.resolver';
+export * from './resolvers/search-page-title.resolver';
