+++ projects/core/src/page/title/facade/resolvers/content-page-title.resolver.ts
@@ -2,9 +2,9 @@ import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { map, filter } from 'rxjs/operators';
 
-import { Page } from '../../../cms/model/page.model';
-import { PageType } from '../../../occ/occ-models/occ.models';
-import { CmsService } from '../../../cms/facade/cms.service';
+import { Page } from '../../../../cms/model/page.model';
+import { PageType } from '../../../../occ/occ-models/occ.models';
+import { CmsService } from '../../../../cms/facade/cms.service';
 import { PageTitleResolver } from './page-title.resolver';
 
 @Injectable({
