+++ projects/core/src/page/title/facade/resolvers/page-title.resolver.ts
@@ -1,4 +1,4 @@
-import { Page } from '../../../cms/model/page.model';
+import { Page } from '../../../../cms/model/page.model';
 import { Observable } from 'rxjs';
 
 export abstract class PageTitleResolver {
