+++ projects/core/src/page/title/facade/resolvers/product-page-title.resolver.ts
