+++ projects/core/src/page/title/facade/resolvers/search-page-title.resolver.ts
